pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grit"
version = "11.0.0"
description = "Grid Toolkit for Grafana"
readme = "README.md"
keywords = ["grafana", "grafanalib", "generator"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
requires-python = ">=3.9, <4"
license = { text = "MIT" }
maintainers = [
    { name = "Hawk", email = "team-sre@hawk.ai" }
]
dependencies = [
    "attrs==21.4.0",
    "pydantic==1.9.1",
    "pydantic-argparse==0.5.0",
    "python-decouple==3.6",
    "python-dotenv==0.20.0",
    "PyYAML==6.0",
    "requests==2.28.0",
    "grafanalib @ git+https://github.com/hawk-ai-aml/grafanalib.git"
]

[tool.setuptools.package-dir]
"" = "src"
